# openapi/openapi.json
{
  "components": {
    "parameters": {},
    "schemas": {
      "CreateProjectRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 3,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "HealthResponse": {
        "additionalProperties": false,
        "properties": {
          "service": {
            "enum": [
              "generated-app"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "ok"
            ],
            "type": "string"
          },
          "version": {
            "enum": [
              1
            ],
            "type": "number"
          }
        },
        "required": [
          "service",
          "status",
          "version"
        ],
        "type": "object"
      },
      "ProblemDetail": {
        "additionalProperties": false,
        "properties": {
          "correlationId": {
            "minLength": 1,
            "type": "string"
          },
          "detail": {
            "minLength": 1,
            "type": "string"
          },
          "instance": {
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "correlationId",
          "detail",
          "instance",
          "status",
          "title",
          "type"
        ],
        "type": "object"
      },
      "ProjectResponse": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 3,
            "type": "string"
          },
          "revision": {
            "exclusiveMinimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "createdAt",
          "id",
          "name",
          "revision"
        ],
        "type": "object"
      },
      "ProtectedFeatureResponse": {
        "additionalProperties": false,
        "properties": {
          "capability": {
            "enum": [
              "protected-insights"
            ],
            "type": "string"
          },
          "projectId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              "available"
            ],
            "type": "string"
          }
        },
        "required": [
          "capability",
          "projectId",
          "status"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Strict current API for an App Builder generated application.",
    "title": "Generated Application API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/health": {
      "get": {
        "parameters": [
          {
            "in": "header",
            "name": "API-Version",
            "required": true,
            "schema": {
              "enum": [
                "1"
              ],
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Generated application API health."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The requested API version is unsupported."
          }
        },
        "tags": [
          "system"
        ]
      }
    },
    "/api/v1/projects": {
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "API-Version",
            "required": true,
            "schema": {
              "enum": [
                "1"
              ],
              "example": "1",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "The project was created transactionally."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The request or API version is invalid."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "A database-backed session is required."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The user is not a member of the requested organization."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "A project with the normalized name already exists."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The transaction failed without exposing internal details."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The identity boundary is not configured."
          }
        },
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{projectId}/protected-feature": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "API-Version",
            "required": true,
            "schema": {
              "enum": [
                "1"
              ],
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtectedFeatureResponse"
                }
              }
            },
            "description": "An owner-only, organization-scoped feature backed by an entitlement."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The project identifier or API version is invalid."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "A database-backed session is required."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The owning organization lacks the required entitlement."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The project does not exist or is concealed by its ownership boundary."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Authorization failed without exposing internal details."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The identity boundary is not configured."
          }
        },
        "tags": [
          "projects"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "webhooks": {}
}