# src/api/generated/models/createProjectRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Generated Application API
 * Strict current API for an App Builder generated application.
 * OpenAPI spec version: 1.0.0
 */

export interface CreateProjectRequest {
  /**
   * @minLength 3
   * @maxLength 100
   */
  name: string;
}