# src/api/generated/models/healthResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Generated Application API
 * Strict current API for an App Builder generated application.
 * OpenAPI spec version: 1.0.0
 */
import type { HealthResponseService } from "./healthResponseService.ts";
import type { HealthResponseStatus } from "./healthResponseStatus.ts";
import type { HealthResponseVersion } from "./healthResponseVersion.ts";

export interface HealthResponse {
  service: HealthResponseService;
  status: HealthResponseStatus;
  version: HealthResponseVersion;
}