# src/api/generated/models/problemDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Generated Application API
 * Strict current API for an App Builder generated application.
 * OpenAPI spec version: 1.0.0
 */

export interface ProblemDetail {
  /** @minLength 1 */
  correlationId: string;
  /** @minLength 1 */
  detail: string;
  /** @minLength 1 */
  instance: string;
  /**
   * @minimum 400
   * @maximum 599
   */
  status: number;
  /** @minLength 1 */
  title: string;
  type: string;
}