# src/api/generated/models/protectedFeatureResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Generated Application API
 * Strict current API for an App Builder generated application.
 * OpenAPI spec version: 1.0.0
 */
import type { ProtectedFeatureResponseCapability } from "./protectedFeatureResponseCapability.ts";
import type { ProtectedFeatureResponseStatus } from "./protectedFeatureResponseStatus.ts";

export interface ProtectedFeatureResponse {
  capability: ProtectedFeatureResponseCapability;
  projectId: string;
  status: ProtectedFeatureResponseStatus;
}